Moved classes to one package
