+++ src/main/java/org/myapps/youtube/commentranker/CommentData.java
@@ -1,4 +1,4 @@
-package comment;
+package org.myapps.youtube.commentranker;
 
 import com.google.api.client.util.DateTime;
