+++ src/main/java/org/myapps/youtube/commentranker/CommentThreadData.java
@@ -1,4 +1,4 @@
-package comment;
+package org.myapps.youtube.commentranker;
 
 import java.util.List;
