+++ src/main/java/org/myapps/youtube/commentranker/Video.java
@@ -1,4 +1,4 @@
-package video;
+package org.myapps.youtube.commentranker;
 
 import com.google.api.client.util.DateTime;
